+++ packages/featureserver/CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Coerce all query parameters to true or false before any function is executed
+
 ## [1.0.6] - 2017-02-01
 ### Fixed
 * Pre es6 compatiblity